raft: move broker_v0 definition into raft subsystem
The broker_v0 is used for serialization compat. This commit moves its
definition into raft which is the only user of the type, rather than
storing it in one of the most used headers in the tree.

broker_v0 may actually not be needed any longer, but that will require a
different type of analysis.
