+++ src/v/model/adl_serde.h
@@ -89,12 +89,6 @@ struct adl<model::broker> {
     model::broker from(iobuf_parser& in);
 };
 
-template<>
-struct adl<model::internal::broker_v0> {
-    void to(iobuf& out, model::internal::broker_v0&& r);
-    model::internal::broker_v0 from(iobuf_parser& in);
-};
-
 // TODO: optimize this transmition with varints
 template<>
 struct adl<model::record> {
