+++ src/v/raft/tests/configuration_serialization_test.cc
@@ -12,6 +12,7 @@
 #include "model/metadata.h"
 #include "model/tests/random_batch.h"
 #include "model/tests/randoms.h"
+#include "raft/broker_compat.h"
 #include "raft/consensus_utils.h"
 #include "raft/group_configuration.h"
 #include "random/generators.h"
